+++ Runtime/CollisionCaster.cs
@@ -10,130 +10,137 @@ namespace TSL.Collisions {
     /// </summary>
     /// <returns>The distance between the object's feet and the closest piece of ground.</returns>
     /// <seealso cref="CollisionCaster.DistanceToGround" />
-    float DistanceToGround(Vector2 center, Vector2 extents);
+    float DistanceToGround(int numCasts = 5);
 
     /// <summary>
-    /// How far the object is from a left-hand wall.
+    /// How far the object is from the closest ceiling.
     /// </summary>
-    /// <returns>The distance between the object's left side and the closest left-hand wall.</returns>
+    /// <returns>The distance between the object and the closest ceiling.</returns>
+    /// <seealso cref="CollisionCaster.DistanceToCeiling" />
+    float DistanceToCeiling(int numCasts = 5);
+
+    /// <summary>
+    /// How far the player is from a left-hand wall.
+    /// </summary>
+    /// <returns>The distance between the player's left side and the closest left-hand wall.</returns>
     /// <seealso cref="CollisionCaster.DistanceToLeftWall" />
-    float DistanceToLeftWall(Vector2 center, Vector2 extents, int checks = 5);
+    float DistanceToLeftWall(int numCasts = 5);
 
     /// <summary>
-    /// How far the object is from a right-hand wall.
+    /// How far the player is from a right-hand wall.
     /// </summary>
-    /// <returns>The distance between the object's right side and the closest right-hand wall.</returns>
+    /// <returns>The distance between the player's right side and the closest right-hand wall.</returns>
     /// <seealso cref="CollisionCaster.DistanceToRightWall" />
-    float DistanceToRightWall(Vector2 center, Vector2 extents, int checks = 5);
+    float DistanceToRightWall(int numCasts = 5);
 
     /// <summary>
     /// Whether or not the top right of the collider is close to a wall.
     /// </summary>
     /// <returns>True if the collider's top right corner is by a wall. False otherwise.</returns>
-    /// <seealso cref="CollisionCaster.IsTopByRightWall"/>
-    float TopDistanceToRightWall(Vector2 center, Vector2 extents);
+    /// <seealso cref="CollisionCaster.TopDistanceToRightWall" />
+    float TopDistanceToRightWall();
 
     /// <summary>
     /// Whether or not the bottom right of the colllider is close to a wall.
     /// </summary>
     /// <returns>True if the collider's bottom right corner is by a wall. False otherwise.</returns>
-    /// <seealso cref="CollisionCaster.BottomDistanceToRightWall"/>
-    float BottomDistanceToRightWall(Vector2 center, Vector2 extents);
+    /// <seealso cref="CollisionCaster.BottomDistanceToRightWall" />
+    float BottomDistanceToRightWall();
 
     /// <summary>
     /// Whether or not the top left of the collider is close to a wall.
     /// </summary>
     /// <returns>True if the collider's top left corner is by a wall. False otherwise.</returns>
-    /// <seealso cref="CollisionCaster.TopDistanceToLeftWall"/>
-    float TopDistanceToLeftWall(Vector2 center, Vector2 extents);
+    /// <seealso cref="CollisionCaster.TopDistanceToLeftWall" />
+    float TopDistanceToLeftWall();
 
     /// <summary>
     /// Whether or not the bottom left of the collider is close to a wall.
     /// </summary>
     /// <returns>True if the collider's bottom left corner is by a wall. False otherwise.</returns>
-    /// <seealso cref="CollisionCaster.BottomDistanceToLeftWall"/>
-    float BottomDistanceToLeftWall(Vector2 center, Vector2 extents);
+    /// <seealso cref="CollisionCaster.BottomDistanceToLeftWall" />
+    float BottomDistanceToLeftWall();
 
     /// <summary>
-    /// How far the object is from the closest wall.
+    /// How far the player is from the closest wall.
     /// </summary>
-    /// <returns>The distance between the object and the closest wall.</returns>
+    /// <returns>The distance between the player and the closest wall.</returns>
     /// <seealso cref="CollisionCaster.DistanceToWall" />
-    float DistanceToWall(Vector2 center, Vector2 extents);
+    float DistanceToWall(int numCasts = 5);
 
     /// <summary>
-    /// How far the object is from the closest ceiling.
-    /// </summary>
-    /// <seealso cref="CollisionCaster.DistanceToCeiling" />
-    /// <returns>The distance between the object and the closest ceiling.</returns>
-    float DistanceToCeiling(Vector2 center, Vector2 size);
-
-    /// <summary>
-    /// Whether or not the object is touching the ground.
+    /// Whether or not the player is touching the ground.
     /// </summary>
     /// <seealso cref="CollisionCaster.IsTouchingGround" />
-    bool IsTouchingGround(Vector2 center, Vector2 size);
+    bool IsTouchingGround();
 
     /// <summary>
-    /// Whether or not the object is touching a left-hand wall.
+    /// Whether or not the player is touching a left-hand wall.
     /// </summary>
     /// <seealso cref="CollisionCaster.IsTouchingLeftWall" />
-    bool IsTouchingLeftWall(Vector2 center, Vector2 size);
+    bool IsTouchingLeftWall();
 
     /// <summary>
-    /// Whether or not the object is touching a right-hand wall.
+    /// Whether or not the player is touching a right-hand wall.
     /// </summary>
     /// <seealso cref="CollisionCaster.IsTouchingRightWall" />
-    bool IsTouchingRightWall(Vector2 center, Vector2 size);
+    bool IsTouchingRightWall();
+
+    /// <summary>
+    /// Whether or not the collider is hanging over a left-hand ledge.
+    /// </summary>
+    /// <returns>True if the bottom left corner of the collider isn't touching
+    /// the ground.</returns>
+    /// <seealso cref="CollisionCaster.IsOverLeftLedge" />
+    bool IsOverLeftLedge();
+
+    /// <summary>
+    /// Whether or not the collider is hanging over a right-hand ledge.
+    /// </summary>
+    /// <returns>True if the bottom right corner of the collider isn't touching
+    /// the ground.</returns>
+    /// <seealso cref="CollisionCaster.IsOverRightLedge" />
+    bool IsOverRightLedge();
 
     /// <summary>
     /// Whether or not the object is touching the ceiling.
     /// </summary>
     /// <seealso cref="CollisionCaster.IsTouchingCeiling" />
-    bool IsTouchingCeiling(Vector2 center, Vector2 size);
+    bool IsTouchingCeiling();
 
     /// <summary>
     /// Whether or not a box will fit in a position one space below where it
     /// currently is.
     /// </summary>
-    /// <param name="center">The center of the box.</param>
-    /// <param name="size">The dimensions of the box.</param>
     /// <returns>Returns true if the box would fit in the space directly below
     /// it's feet.</returns>
     /// <seealso cref="CollisionCaster.FitsDown" />
-    bool FitsDown(Vector2 center, Vector2 size, out Collider2D[] hits);
+    bool FitsDown(out Collider2D[] hits);
 
     /// <summary>
     /// Whether or not a box will fit in a position one space above where it
     /// currently is.
     /// </summary>
-    /// <param name="center">The center of the box.</param>
-    /// <param name="size">The dimensions of the box.</param>
     /// <returns>Returns true if the box would fit in the space directly above
     /// it's top.</returns>
     /// <seealso cref="CollisionCaster.FitsUp" />
-    bool FitsUp(Vector2 center, Vector2 size, out Collider2D[] hits);
+    bool FitsUp(out Collider2D[] hits);
 
     /// <summary>
     /// Whether or not a box will fit in a position one space to the left of where it
     /// currently is.
     /// </summary>
-    /// <param name="center">The center of the box.</param>
-    /// <param name="size">The dimensions of the box.</param>
     /// <returns>Returns true if the box would fit in the space directly to its left.</returns>
     /// <seealso cref="CollisionCaster.FitsLeft" />
-    bool FitsLeft(Vector2 center, Vector2 size, out Collider2D[] hits);
+    bool FitsLeft(out Collider2D[] hits);
 
     /// <summary>
     /// Whether or not a box will fit in a position one space to the right of where it
     /// currently is.
     /// </summary>
-    /// <param name="center">The center of the box.</param>
-    /// <param name="size">The dimensions of the box.</param>
     /// <returns>Returns true if the box would fit in the space directly to its right.</returns>
     /// <seealso cref="CollisionCaster.FitsRight" />
-    bool FitsRight(Vector2 center, Vector2 size, out Collider2D[] hits);
-
+    bool FitsRight(out Collider2D[] hits);
 
     /// <summary>
     /// Whether or not a box will fit in a position one space to the right of where it
@@ -144,19 +151,7 @@ namespace TSL.Collisions {
     /// <param name="direction">The direction to check</param>
     /// <returns>Returns true if the box would fit in the space directly to its right.</returns>
     /// <seealso cref="CollisionCaster.FitsInDirection" />
-    bool FitsInDirection(Vector2 center, Vector2 size, Vector2 direction, out Collider2D[] hits);
-    
-
-    /// <summary>
-    /// Whether or not a collision can be considered a valid "ground" collision.
-    /// </summary>
-    /// <param name="collider">The collider of the other object (i.e., not the
-    /// player's collider).</param>
-    /// <param name="hitNormal">The normal for the collision.</param>
-    /// <param name="checkNormal">The normal expected.</param>
-    /// <returns>True if this is a valid "ground" hit. False otherwise.</returns>
-    /// <seealso cref="CollisionCaster.IsHit" />
-    bool IsHit(Collider2D collider, Vector2? hitNormal = null, Vector2? checkNormal = null);
+    bool FitsInDirection(Vector2 direction, out Collider2D[] hits);
   }
   #endregion
 
@@ -169,41 +164,47 @@ namespace TSL.Collisions {
     private float colliderWidth = 0.25f;
 
     /// <summary>
-    /// The vertical and horizontal difference between the player's collider and the box cast.
+    /// The vertical and horizontal difference between the collider and the box cast.
     /// </summary>
     private float boxCastMargin = 0f;
 
     /// <summary>
-    /// Layer mask that prevents collisions with anything aside from things on the ground layer.
+    /// Layer mask that prevents collisions with anything aside from things on the given layer(s).
     /// </summary>
     private LayerMask layerMask;
 
     /// <summary>
-    /// The collider this collision component is attached to.
+    /// The list of tags the caster should check when checking for a collision.
+    /// e.g., if the list is ["Ground"], then only colliders tagged attached to
+    /// an object with the tage "Ground" will be considered for collisions.
     /// </summary>
-    private Collider2D parentCollider;
+    private List<string> tags;
+
+    /// <summary>
+    /// The collider this collision checker is for.
+    /// </summary>
+    private Collider2D parent;
     #endregion
 
-
     #region Constructors
-    public CollisionCaster() {
-      layerMask = LayerMask.GetMask("Foreground");
+    public CollisionCaster(Collider2D parent) {
+      this.parent = parent;
+      layerMask = LayerMask.GetMask("Default");
+      this.tags = new List<string>();
     }
 
-    public CollisionCaster(Collider2D parent) : this() {
-      parentCollider = parent;
+    public CollisionCaster(Collider2D parent, string[] layerNames, string[] tags) {
+      this.parent = parent;
+      layerMask = LayerMask.GetMask(layerNames);
+      this.tags = new List<string>(tags);
     }
 
-    public CollisionCaster(float colliderWidth, float boxCastMargin, string[] layerNames) {
+    public CollisionCaster(Collider2D parent, float colliderWidth, float boxCastMargin, string[] layerNames) {
+      this.parent = parent;
       this.colliderWidth = colliderWidth;
       this.boxCastMargin = boxCastMargin;
       layerMask = LayerMask.GetMask(layerNames);
     }
-
-    public CollisionCaster(string[] layerNames) {
-      layerMask = LayerMask.GetMask(layerNames);
-    }
-
     #endregion
 
     #region Distance Checking
@@ -211,34 +212,46 @@ namespace TSL.Collisions {
     /// How far the player is from the ground.
     /// </summary>
     /// <returns>The distance between the player's feet and the closest piece of ground.</returns>
-    public float DistanceToGround(Vector2 center, Vector2 extents) {
-      return CheckDistance(Vector2.down, center, extents, 3);
-    }
+    public float DistanceToGround(int numCasts = 5) => CheckDistance(
+      Vector2.down, 
+      parent.bounds.center,
+      parent.bounds.size,
+      numCasts
+    );
 
 
     /// <summary>
     /// How far the object is from the closest ceiling.
     /// </summary>
     /// <returns>The distance between the object and the closest ceiling.</returns>
-    public float DistanceToCeiling(Vector2 center, Vector2 extents) {
-      return CheckDistance(Vector2.up, center, extents, 3);
-    }
+    public float DistanceToCeiling(int numCasts = 5) => CheckDistance(
+      Vector2.up,
+      parent.bounds.center,
+      parent.bounds.size,
+      numCasts
+    );
 
     /// <summary>
     /// How far the player is from a left-hand wall.
     /// </summary>
     /// <returns>The distance between the player's left side and the closest left-hand wall.</returns>
-    public float DistanceToLeftWall(Vector2 center, Vector2 extents, int checks = 5) {
-      return CheckDistance(Vector2.left, center, extents);
-    }
+    public float DistanceToLeftWall(int numCasts = 5) => CheckDistance(
+      Vector2.left,
+      parent.bounds.center,
+      parent.bounds.size,
+      numCasts
+    );
 
     /// <summary>
     /// How far the player is from a right-hand wall.
     /// </summary>
     /// <returns>The distance between the player's right side and the closest right-hand wall.</returns>
-    public float DistanceToRightWall(Vector2 center, Vector2 extents, int checks = 5) {
-      return CheckDistance(Vector2.right, center, extents);
-    }
+    public float DistanceToRightWall(int numCasts = 5) => CheckDistance(
+      Vector2.right,
+      parent.bounds.center,
+      parent.bounds.size,
+      numCasts
+    );
 
     private float CheckDistance(Vector2 direction, Vector2 center, Vector2 extents, int checks = 5) {
       // Ready to get mathy, bro?
@@ -279,8 +292,9 @@ namespace TSL.Collisions {
     /// Whether or not the top right of the collider is close to a wall.
     /// </summary>
     /// <returns>True if the collider's top right corner is by a wall. False otherwise.</returns>
-    public float TopDistanceToRightWall(Vector2 center, Vector2 extents) {
-      Vector2 startTopRight = center + new Vector2(extents.x, extents.y);
+    public float TopDistanceToRightWall() {
+      Vector2 center = new Vector2(parent.bounds.center.x, parent.bounds.center.y);
+      Vector2 startTopRight = center + new Vector2(parent.bounds.extents.x, parent.bounds.extents.y);
       RaycastHit2D hit = Physics2D.Raycast(startTopRight, Vector2.right, float.PositiveInfinity, layerMask);
       return hit.distance;
     }
@@ -289,8 +303,9 @@ namespace TSL.Collisions {
     /// Whether or not the bottom right of the colllider is close to a wall.
     /// </summary>
     /// <returns>True if the collider's bottom right corner is by a wall. False otherwise.</returns>
-    public float BottomDistanceToRightWall(Vector2 center, Vector2 extents) {
-      Vector2 startBottomRight = center + new Vector2(extents.x, -extents.y);
+    public float BottomDistanceToRightWall() {
+      Vector2 center = new Vector2(parent.bounds.center.x, parent.bounds.center.y);
+      Vector2 startBottomRight = center + new Vector2(parent.bounds.extents.x, -parent.bounds.extents.y);
       RaycastHit2D hit = Physics2D.Raycast(startBottomRight, Vector2.right, float.PositiveInfinity, layerMask);
       return hit.distance;
     }
@@ -299,8 +314,9 @@ namespace TSL.Collisions {
     /// Whether or not the top left of the collider is close to a wall.
     /// </summary>
     /// <returns>True if the collider's top left corner is by a wall. False otherwise.</returns>
-    public float TopDistanceToLeftWall(Vector2 center, Vector2 extents) {
-      Vector2 startTopLeft = center + new Vector2(-(extents.x), extents.y);
+    public float TopDistanceToLeftWall() {
+      Vector2 center = new Vector2(parent.bounds.center.x, parent.bounds.center.y);
+      Vector2 startTopLeft = center + new Vector2(-parent.bounds.extents.x, parent.bounds.extents.y);
       RaycastHit2D hit = Physics2D.Raycast(startTopLeft, Vector2.left, float.PositiveInfinity, layerMask);
       return hit.distance;
     }
@@ -309,8 +325,9 @@ namespace TSL.Collisions {
     /// Whether or not the bottom left of the collider is close to a wall.
     /// </summary>
     /// <returns>True if the collider's bottom left corner is by a wall. False otherwise.</returns>
-    public float BottomDistanceToLeftWall(Vector2 center, Vector2 extents) {
-      Vector2 startBottomLeft = center + new Vector2(-extents.x, -extents.y);
+    public float BottomDistanceToLeftWall() {
+      Vector2 center = new Vector2(parent.bounds.center.x, parent.bounds.center.y);
+      Vector2 startBottomLeft = center + new Vector2(-parent.bounds.extents.x, -parent.bounds.extents.y);
       RaycastHit2D hit = Physics2D.Raycast(startBottomLeft, Vector2.left, float.PositiveInfinity, layerMask);
       return hit.distance;
     }
@@ -319,10 +336,10 @@ namespace TSL.Collisions {
     /// How far the player is from the closest wall.
     /// </summary>
     /// <returns>The distance between the player and the closest wall.</returns>
-    public float DistanceToWall(Vector2 center, Vector2 extents) {
+    public float DistanceToWall(int numCasts = 5) {
       return Mathf.Min(new float[] {
-        DistanceToLeftWall(center, extents),
-        DistanceToRightWall(center, extents)
+        DistanceToLeftWall(numCasts),
+        DistanceToRightWall(numCasts)
       });
     }
     #endregion
@@ -331,7 +348,8 @@ namespace TSL.Collisions {
     /// <summary>
     /// Whether or not the player is touching the ground.
     /// </summary>
-    public bool IsTouchingGround(Vector2 center, Vector2 size) {
+    public bool IsTouchingGround() => IsTouchingGround_Inner(parent.bounds.center, parent.bounds.size);
+    private bool IsTouchingGround_Inner(Vector2 center, Vector2 size) {
       Vector2 boxCast = size - new Vector2(boxCastMargin, 0);
 
       Vector2 startLeft = center - new Vector2(center.x, center.y-size.y/2);
@@ -358,14 +376,8 @@ namespace TSL.Collisions {
     /// <summary>
     /// Whether or not the player is touching a left-hand wall.
     /// </summary>
-    public bool IsTouchingLeftWall() {
-      return IsTouchingLeftWall(parentCollider.bounds.center, parentCollider.bounds.size);
-    }
-
-    /// <summary>
-    /// Whether or not the player is touching a left-hand wall.
-    /// </summary>
-    public bool IsTouchingLeftWall(Vector2 center, Vector2 size) {
+    public bool IsTouchingLeftWall() => IsTouchingLeftWall_Inner(parent.bounds.center, parent.bounds.size);
+    private bool IsTouchingLeftWall_Inner(Vector2 center, Vector2 size) {
       Vector2 boxCast = size - new Vector2(0, boxCastMargin);
 
       RaycastHit2D[] hits = Physics2D.BoxCastAll(
@@ -383,14 +395,8 @@ namespace TSL.Collisions {
     /// <summary>
     /// Whether or not the player is touching a right-hand wall.
     /// </summary>
-    public bool IsTouchingRightWall() {
-      return IsTouchingRightWall(parentCollider.bounds.center, parentCollider.bounds.size);
-    }
-
-    /// <summary>
-    /// Whether or not the player is touching a right-hand wall.
-    /// </summary>
-    public bool IsTouchingRightWall(Vector2 center, Vector2 size) {
+    public bool IsTouchingRightWall() => IsTouchingRightWall_Inner(parent.bounds.center, parent.bounds.size);
+    private bool IsTouchingRightWall_Inner(Vector2 center, Vector2 size) {
       Vector2 boxCast = size - new Vector2(0, boxCastMargin);
 
 
@@ -411,16 +417,8 @@ namespace TSL.Collisions {
     /// </summary>
     /// <returns>True if the bottom left corner of the collider isn't touching
     /// the ground.</returns>
-    public bool IsOverLeftLedge() {
-      return IsOverLeftLedge(parentCollider.bounds.center, parentCollider.bounds.size);
-    }
-
-    /// <summary>
-    /// Whether or not the collider is hanging over a left-hand ledge.
-    /// </summary>
-    /// <returns>True if the bottom left corner of the collider isn't touching
-    /// the ground.</returns>
-    public bool IsOverLeftLedge(Vector2 center, Vector2 size) {
+    public bool IsOverLeftLedge() => IsOverLeftLedge_Inner(parent.bounds.center, parent.bounds.size);
+    private bool IsOverLeftLedge_Inner(Vector2 center, Vector2 size) {
       Vector2 bottomLeftCorner = new Vector2(
         center.x-(size.x/2),
         center.y-(size.y/2)
@@ -441,16 +439,8 @@ namespace TSL.Collisions {
     /// </summary>
     /// <returns>True if the bottom right corner of the collider isn't touching
     /// the ground.</returns>
-    public bool IsOverRightLedge() {
-      return IsOverRightLedge(parentCollider.bounds.center, parentCollider.bounds.size);
-    }
-
-    /// <summary>
-    /// Whether or not the collider is hanging over a right-hand ledge.
-    /// </summary>
-    /// <returns>True if the bottom right corner of the collider isn't touching
-    /// the ground.</returns>
-    public bool IsOverRightLedge(Vector2 center, Vector2 size) {
+    public bool IsOverRightLedge() => IsOverRightLedge_Inner(parent.bounds.center, parent.bounds.size);
+    private bool IsOverRightLedge_Inner(Vector2 center, Vector2 size) {
       Vector2 bottomLeftCorner = new Vector2(
         center.x+(size.x/2),
         center.y+(size.y/2)
@@ -469,7 +459,8 @@ namespace TSL.Collisions {
     /// <summary>
     /// Whether or not the object is touching the ceiling.
     /// </summary>
-    public bool IsTouchingCeiling(Vector2 center, Vector2 size) {
+    public bool IsTouchingCeiling() => IsTouchingCeiling_Inner(parent.bounds.center, parent.bounds.size);
+    private bool IsTouchingCeiling_Inner(Vector2 center, Vector2 size) {
       Vector2 boxCast = size - new Vector2(0, boxCastMargin);
 
       RaycastHit2D[] hits = Physics2D.BoxCastAll(
@@ -491,7 +482,7 @@ namespace TSL.Collisions {
     /// <param name="hits">The list of RaycastHits</param>
     /// <param name="normalDirection">The normal of the direction to check hits against.</param>
     /// <returns>Whether or not there are any ground contacts in the desired direction.</returns>
-    public bool AnyHits(RaycastHit2D[] hits, Vector2 normalDirection) {
+    private bool AnyHits(RaycastHit2D[] hits, Vector2 normalDirection) {
       for (int i = 0; i < hits.Length; i++) {
         if (IsHit(hits[i].collider, hits[i].normal, normalDirection)) {
           return true;
@@ -509,7 +500,7 @@ namespace TSL.Collisions {
     /// <param name="hitNormal">The normal for the collision.</param>
     /// <param name="checkNormal">The normal expected.</param>
     /// <returns>True if this is a valid "ground" hit. False otherwise.</returns>
-    public bool IsHit(Collider2D collider, Vector2? hitNormal = null, Vector2? checkNormal = null) {
+    private bool IsHit(Collider2D collider, Vector2? hitNormal = null, Vector2? checkNormal = null) {
       if (collider == null) {
         return false;
       }
@@ -527,14 +518,24 @@ namespace TSL.Collisions {
     /// </summary>
     /// <param name="collider">The collider</param>
     /// <returns>True if this is a valid "ground" overlap.</returns>
-    public bool IsOverlap(Collider2D collider) {
+    private bool IsOverlap(Collider2D collider) {
       if (collider == null) {
         return false;
       }
 
-      return (parentCollider == null || !Physics2D.GetIgnoreCollision(collider, parentCollider)) && // Collision isn't purposefully ignored,
-        collider.CompareTag("Ground") &&                                                            // Collider is a ground object,
-        !collider.isTrigger;                                                                        // & Collider isn't a trigger.
+      return (parent == null || !Physics2D.GetIgnoreCollision(collider, parent)) && // Collision isn't purposefully ignored,
+        IsTagged(collider) && // Collider is tagged with something we care about
+        !collider.isTrigger; // & Collider isn't a trigger.
+    }
+
+    private bool IsTagged(Collider2D collider) {
+      foreach (string tag in tags) {
+        if (!collider.CompareTag(tag)) {
+          return false;
+        }
+      }
+
+      return true;
     }
 
 
@@ -542,11 +543,10 @@ namespace TSL.Collisions {
     /// Whether or not a box will fit in a position one space below where it
     /// currently is.
     /// </summary>
-    /// <param name="center">The center of the box.</param>
-    /// <param name="size">The dimensions of the box.</param>
     /// <returns>Returns true if the box would fit in the space directly below
     /// it's feet.</returns>
-    public bool FitsDown(Vector2 center, Vector2 size, out Collider2D[] hits) {
+    public bool FitsDown(out Collider2D[] hits) => FitsDown(parent.bounds.center, parent.bounds.size, out hits);
+    private bool FitsDown(Vector2 center, Vector2 size, out Collider2D[] hits) {
       Vector2 newPosition = center - new Vector2(0, size.y);
       Collider2D[] results = Physics2D.OverlapBoxAll(newPosition, size, 0, layerMask);
 
@@ -565,11 +565,10 @@ namespace TSL.Collisions {
     /// Whether or not a box will fit in a position one space above where it
     /// currently is.
     /// </summary>
-    /// <param name="center">The center of the box.</param>
-    /// <param name="size">The dimensions of the box.</param>
     /// <returns>Returns true if the box would fit in the space directly above
     /// it's top.</returns>
-    public bool FitsUp(Vector2 center, Vector2 size, out Collider2D[] hits) {
+    public bool FitsUp(out Collider2D[] hits) => FitsUp(parent.bounds.center, parent.bounds.size, out hits);
+    private bool FitsUp(Vector2 center, Vector2 size, out Collider2D[] hits) {
       Vector2 newPosition = center + new Vector2(0, size.y);
       hits = Physics2D.OverlapBoxAll(newPosition, size, 0, layerMask);
       return hits.Length == 0;
@@ -579,10 +578,9 @@ namespace TSL.Collisions {
     /// Whether or not a box will fit in a position one space to the left of where it
     /// currently is.
     /// </summary>
-    /// <param name="center">The center of the box.</param>
-    /// <param name="size">The dimensions of the box.</param>
     /// <returns>Returns true if the box would fit in the space directly to its left.</returns>
-    public bool FitsLeft(Vector2 center, Vector2 size, out Collider2D[] hits) {
+    public bool FitsLeft(out Collider2D[] hits) => FitsLeft(parent.bounds.center, parent.bounds.size, out hits);
+    private bool FitsLeft(Vector2 center, Vector2 size, out Collider2D[] hits) {
       Vector2 newPosition = center - new Vector2(size.x, 0);
       hits = Physics2D.OverlapBoxAll(newPosition, size, 0, layerMask);
       return hits.Length == 0;
@@ -592,10 +590,9 @@ namespace TSL.Collisions {
     /// Whether or not a box will fit in a position one space to the right of where it
     /// currently is.
     /// </summary>
-    /// <param name="center">The center of the box.</param>
-    /// <param name="size">The dimensions of the box.</param>
     /// <returns>Returns true if the box would fit in the space directly to its right.</returns>
-    public bool FitsRight(Vector2 center, Vector2 size, out Collider2D[] hits) {
+    public bool FitsRight(out Collider2D[] hits) => FitsRight(parent.bounds.center, parent.bounds.size, out hits);
+    private bool FitsRight(Vector2 center, Vector2 size, out Collider2D[] hits) {
       Vector2 newPosition = center + new Vector2(size.x, 0);
       hits = Physics2D.OverlapBoxAll(newPosition, size, 0, layerMask);
       return hits.Length == 0;
@@ -609,39 +606,12 @@ namespace TSL.Collisions {
     /// <param name="size">The dimensions of the box.</param>
     /// <param name="direction">The direction to check</param>
     /// <returns>Returns true if the box would fit in the space directly to its right.</returns>
-    /// <seealso cref="CollisionCaster.FitsInDirection" />
-    public bool FitsInDirection(Vector2 center, Vector2 size, Vector2 direction, out Collider2D[] hits) {
+    public bool FitsInDirection(Vector2 direction, out Collider2D[] hits) => FitsInDirection(parent.bounds.center, parent.bounds.size, direction, out hits);
+    private bool FitsInDirection(Vector2 center, Vector2 size, Vector2 direction, out Collider2D[] hits) {
       Vector2 newPosition = center + size*direction.normalized;
       hits = Physics2D.OverlapBoxAll(newPosition, size, 0, layerMask);
       return hits.Length == 0;
     }
-
-    // /// <summary>
-    // /// Whether or not the collider is touching a specific collider.
-    // /// </summary>
-    // /// <param name="col">The colllider to check.</param>
-    // /// <returns>True if the colliders overlap.</returns>
-    // public bool IsTouching(Collider2D col) {
-
-    // }
-    #endregion
-
-
-    #region Getters/Setters
-
-    public void SetLayerMask(string[] layerNames) {
-      layerMask = LayerMask.GetMask(layerNames);
-    }
-
-    public void SetBoxCastMargin(float margin) {
-      boxCastMargin = margin;
-    }
-
-    public void SetColliderWidth(float width) {
-      colliderWidth = width;
-    }
-
     #endregion
   }
-
 }